Write rebalance function and refactor insert
rebalance traverses the Binary Search Tree in an inOrder Manner,
inOrder gets passed a callback that appends the node's data to an empty
array resulting in a sorted array of the bst. Finally this sorted array
gets passed to the builTree function to build a balanced tree again if the bst
is not balanced.

Now insert can handle values that are already in the bst. If insert gets passed
a value that's already present in the bst, it simply returns.